Not allow delete subscription with bookings

+++ src/GymManagement.Application/Subscriptions/Commands/DeleteSubscription/DeleteSubscriptionCommandHandler.cs
@@ -13,16 +13,19 @@ namespace GymManagement.Application.Subscriptions.Commands.DeleteSubscription;
 public class DeleteSubscriptionCommandHandler : IRequestHandler<DeleteSubscriptionCommand, ErrorOr<Unit>>
 {
     private readonly ISubscriptionsRepository _subscriptionsRepository;
+    public readonly IBookingsRepository _bookingsRepository;
     private readonly IMembersRepository _membersRepository;
     private readonly IUnitOfWork _unitOfWork;
-                                           
+
     public DeleteSubscriptionCommandHandler(IUnitOfWork unitOfWork,
                                             ISubscriptionsRepository subscriptionsRepository,
-                                            IMembersRepository membersRepository)
+                                            IMembersRepository membersRepository,
+                                            IBookingsRepository bookingsRepository)
     {
         _subscriptionsRepository = subscriptionsRepository;
         _unitOfWork = unitOfWork;
         _membersRepository = membersRepository;
+        _bookingsRepository = bookingsRepository;
     }
 
     public async Task<ErrorOr<Unit>> Handle(DeleteSubscriptionCommand command, CancellationToken cancellationToken = default)
@@ -34,6 +37,11 @@ public class DeleteSubscriptionCommandHandler : IRequestHandler<DeleteSubscripti
         if (subscription.IsActive)
             return SubscriptionErrors.CantDeleteActiveSubscription();
 
+        var bookings = await _bookingsRepository.ListByMemberAsync(subscription.MemberId);
+
+        if (bookings is not null && bookings.Any())
+            return SubscriptionErrors.CantDeleteSubscriptionWithBookings(subscription.Id);
+        
         await _subscriptionsRepository.RemoveAsync(subscription, cancellationToken); 
         await _unitOfWork.CommitChangesAsync(cancellationToken);
